# 32bit genotypes