… c/tests/test_genotypes.c
… c/tests/test_trees.c
… python/tests/test_genotypes.py