Fixes #463

WIP, getting some odd, unrelated test failures. Trying on CI.