@jeromekelleher 2022-01-27
Might be worth doing a check perf check to see what promoting up to 32 bit before going through the painful stuff? Where should be a quick s/i8/i32 change you could do globally that would be easy.