@benjeffery 2022-01-28
> Might be worth doing a check perf check to see what promoting up to 32 bit before going through the painful stuff? Where should be a quick s/i8/i32 change you could do globally that would be easy.

In tracing all the types we also have to change to 32 in map_mutations so doing the lot and then testing that and genotypes for perf.