@bhaller 2022-02-03
If I understand what's being proposed, this should be fine.  We only use vargen_t for crosschecking the tskit data structures against SLiM (i.e., in debug mode) and when loading a .trees file from disk, I believe.  In all places where we currently call tsk_vargen_init() we pass TSK_16_BIT_GENOTYPES; we don't use 8-bit.  Is 16-bit still going to be supported, or is it also going away in favor of 32-bit?  Is the TSK_16_BIT_GENOTYPES flag going to be deprecated, or honored, or just ignored?  Anyhow, I could dig into the diffs here to answer my own questions, but this really ought to be no problem for us, if I'm not totally misunderstanding what is at issue.