@benjeffery 2022-02-04
Some testing:
![Screenshot from 2022-02-04 10-59-47](https://user-images.githubusercontent.com/8552/152517935-1e30bdcd-6259-472c-a733-32571f28e3d7.png)


Python:
```
for v in ts.variants():
    v.genotypes
```
This branch:  129s
Main: 134s

C:
```
    ret = tsk_treeseq_load(&ts, "perf.ts", 0);
    CU_ASSERT_EQUAL_FATAL(ret, 0);
    ret = tsk_vargen_init(&vargen, &ts, samples, 2, NULL, 0);
    CU_ASSERT_EQUAL_FATAL(ret, 0);
    while (tsk_vargen_next(&vargen, &var) == 1){
    };
```
This branch: 312s
Main: 312s

What I don't understand is why the python is so much faster. Must be a mistake right? Looking into it.