@benjeffery 2022-02-04
Ok think I have sorted my issues out.
Looks like the 32bit genotypes branch is faster by 2-3%.