@benjeffery 2022-02-04
No noticeable impact on ram, as you would expect for iterating. Just needs a changelog.