@bhaller 2022-02-04
> Thanks for the questions @bhaller - under the current proposal the 16bit option will be removed and the flag will be ignored, hence why we want to get this sorted for C API 1.0.

OK, good to know.  Full steam ahead; will be a trivial fix when we decide to merge the next tskit version, I think.  Thanks.