@benjeffery 2022-02-07
@jeromekelleher What sort of info did you want from `perf`?

Here's the summary:
`main`:
```
         43,598.48 msec task-clock                #    1.000 CPUs utilized          
               594      context-switches          #    0.014 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
            32,826      page-faults               #    0.753 K/sec                  
   198,588,349,911      cycles                    #    4.555 GHz                    
   215,228,429,541      instructions              #    1.08  insn per cycle         
    19,162,813,346      branches                  #  439.529 M/sec                  
           778,793      branch-misses             #    0.00% of all branches        

      43.606464362 seconds time elapsed

      43.554717000 seconds user
       0.044002000 seconds sys
```
`branch`:
```
         45,034.81 msec task-clock                #    1.000 CPUs utilized          
               130      context-switches          #    0.003 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
            33,219      page-faults               #    0.738 K/sec                  
   205,385,182,502      cycles                    #    4.561 GHz                    
   234,245,790,040      instructions              #    1.14  insn per cycle         
    19,162,553,042      branches                  #  425.505 M/sec                  
           795,300      branch-misses             #    0.00% of all branches        

      45.036363412 seconds time elapsed

      44.991126000 seconds user
       0.043999000 seconds sys
```
The flame graph from both is almost 100% `tsk_variant_update_site`

![Screenshot from 2022-02-07 12-46-59](https://user-images.githubusercontent.com/8552/152790812-6d57c308-4e18-48f1-97e7-b3b593f6bdc9.png)