@jeromekelleher 2022-02-07
Looks great! The first ~10 lines of the ``perf report`` screen for both is the most informative I think.