@jeromekelleher 2022-02-07
OK, so there's a very slight increase in the time spent in ``tsk_variant_update_genotypes_sample_list``, but it's within the range we'd expect for statistical noise. 

I'm sold then, let's make the change!